requirements.txt
chromadb==0.5.2
langchain-text-splitters==0.2.1
langchain_community==0.2.5
pypdf==4.2.0
PyMuPDF==1.24.5
pillow==10.3.0
langchain_google_genai==1.0.6
